[Oracle] Sources Fix Fix mismatch in types when sources are passed as a list of strings instead of a list of dictionaries
